+++ internal/app/http_auth.go
@@ -1,11 +1,8 @@
 package app
 
 import (
-	"context"
 	"crypto/sha256"
-	"encoding/json"
 	"fmt"
-	"io"
 	"net"
 	"net/http"
 	"net/url"
@@ -18,7 +15,6 @@ import (
 	"github.com/golang-jwt/jwt/v5"
 	"github.com/leighmacdonald/gbans/internal/consts"
 	"github.com/leighmacdonald/gbans/internal/store"
-	"github.com/leighmacdonald/gbans/pkg/util"
 	"github.com/leighmacdonald/steamid/v3/steamid"
 	"github.com/pkg/errors"
 	"github.com/yohcop/openid-go"
@@ -111,221 +107,6 @@ func referral(ctx *gin.Context) string {
 	}
 
 	return referralURL
-}
-
-func onOAuthDiscordCallback(app *App) gin.HandlerFunc {
-	type accessTokenResp struct {
-		AccessToken  string `json:"access_token"`
-		ExpiresIn    int    `json:"expires_in"`
-		RefreshToken string `json:"refresh_token"`
-		Scope        string `json:"scope"`
-		TokenType    string `json:"token_type"`
-	}
-
-	type discordUserDetail struct {
-		ID               string      `json:"id"`
-		Username         string      `json:"username"`
-		Avatar           string      `json:"avatar"`
-		AvatarDecoration interface{} `json:"avatar_decoration"`
-		Discriminator    string      `json:"discriminator"`
-		PublicFlags      int         `json:"public_flags"`
-		Flags            int         `json:"flags"`
-		Banner           interface{} `json:"banner"`
-		BannerColor      interface{} `json:"banner_color"`
-		AccentColor      interface{} `json:"accent_color"`
-		Locale           string      `json:"locale"`
-		MfaEnabled       bool        `json:"mfa_enabled"`
-		PremiumType      int         `json:"premium_type"`
-	}
-
-	log := app.log.Named(runtime.FuncForPC(make([]uintptr, 10)[0]).Name())
-	client := util.NewHTTPClient()
-
-	fetchDiscordID := func(ctx context.Context, accessToken string) (string, error) {
-		req, errReq := http.NewRequestWithContext(ctx, http.MethodGet, "https://discord.com/api/users/@me", nil)
-		if errReq != nil {
-			return "", errors.Wrap(errReq, "Failed to create new request")
-		}
-
-		req.Header.Add("Authorization", fmt.Sprintf("Bearer %s", accessToken))
-		resp, errResp := client.Do(req)
-
-		if errResp != nil {
-			return "", errors.Wrap(errResp, "Failed to perform http request")
-		}
-
-		defer func() {
-			_ = resp.Body.Close()
-		}()
-
-		body, errBody := io.ReadAll(resp.Body)
-		if errBody != nil {
-			return "", errors.Wrap(errBody, "Failed to read response body")
-		}
-
-		var details discordUserDetail
-		if errJSON := json.Unmarshal(body, &details); errJSON != nil {
-			return "", errors.Wrap(errJSON, "Failed to unmarshal response")
-		}
-
-		return details.ID, nil
-	}
-
-	fetchToken := func(ctx context.Context, code string) (string, error) {
-		// v, _ := go_oauth_pkce_code_verifier.CreateCodeVerifierFromBytes([]byte(code))
-		form := url.Values{}
-		form.Set("client_id", app.conf.Discord.AppID)
-		form.Set("client_secret", app.conf.Discord.AppSecret)
-		form.Set("redirect_uri", app.ExtURLRaw("/login/discord"))
-		form.Set("code", code)
-		form.Set("grant_type", "authorization_code")
-		// form.Set("state", state.String())
-		form.Set("scope", "identify")
-		req, errReq := http.NewRequestWithContext(ctx, http.MethodPost, "https://discord.com/api/oauth2/token", strings.NewReader(form.Encode()))
-
-		if errReq != nil {
-			return "", errors.Wrap(errReq, "Failed to create new request")
-		}
-
-		req.Header.Add("Content-Type", "application/x-www-form-urlencoded")
-
-		resp, errResp := client.Do(req)
-		if errResp != nil {
-			return "", errors.Wrap(errResp, "Failed to perform http request")
-		}
-
-		defer func() {
-			_ = resp.Body.Close()
-		}()
-
-		body, errBody := io.ReadAll(resp.Body)
-		if errBody != nil {
-			return "", errors.Wrap(errBody, "Failed to read response body")
-		}
-
-		var atr accessTokenResp
-		if errJSON := json.Unmarshal(body, &atr); errJSON != nil {
-			return "", errors.Wrap(errJSON, "Failed to decode response body")
-		}
-
-		if atr.AccessToken == "" {
-			return "", errors.New("Empty token returned")
-		}
-
-		return atr.AccessToken, nil
-	}
-
-	return func(ctx *gin.Context) {
-		code := ctx.Query("code")
-		if code == "" {
-			responseErr(ctx, http.StatusBadRequest, nil)
-			log.Error("Failed to get code from query")
-
-			return
-		}
-
-		token, errToken := fetchToken(ctx, code)
-		if errToken != nil {
-			responseErr(ctx, http.StatusBadRequest, nil)
-			log.Error("Failed to fetch token", zap.Error(errToken))
-
-			return
-		}
-
-		discordID, errID := fetchDiscordID(ctx, token)
-		if errID != nil {
-			responseErr(ctx, http.StatusBadRequest, nil)
-			log.Error("Failed to fetch discord ID", zap.Error(errID))
-
-			return
-		}
-
-		if discordID == "" {
-			responseErr(ctx, http.StatusInternalServerError, nil)
-			log.Error("Empty discord id received")
-
-			return
-		}
-
-		var discordPerson store.Person
-		if errDp := app.db.GetPersonByDiscordID(ctx, discordID, &discordPerson); errDp != nil {
-			if !errors.Is(errDp, store.ErrNoResult) {
-				responseErr(ctx, http.StatusInternalServerError, nil)
-
-				return
-			}
-		}
-
-		if discordPerson.DiscordID != "" {
-			responseErr(ctx, http.StatusConflict, nil)
-			log.Error("Failed to update persons discord id")
-
-			return
-		}
-
-		sid := currentUserProfile(ctx).SteamID
-
-		var person store.Person
-		if errPerson := app.PersonBySID(ctx, sid, &person); errPerson != nil {
-			responseErr(ctx, http.StatusInternalServerError, nil)
-
-			return
-		}
-
-		person.DiscordID = discordID
-
-		if errSave := app.db.SavePerson(ctx, &person); errSave != nil {
-			responseErr(ctx, http.StatusInternalServerError, nil)
-
-			return
-		}
-
-		ctx.JSON(http.StatusOK, nil)
-
-		log.Info("Discord account linked successfully",
-			zap.String("discord_id", discordID), zap.Int64("sid64", sid.Int64()))
-	}
-}
-
-func onAPILogout(app *App) gin.HandlerFunc {
-	log := app.log.Named(runtime.FuncForPC(make([]uintptr, 10)[0]).Name())
-
-	return func(ctx *gin.Context) {
-		fingerprint, errCookie := ctx.Cookie(fingerprintCookieName)
-		if errCookie != nil {
-			responseErr(ctx, http.StatusInternalServerError, nil)
-
-			return
-		}
-
-		parsedExternal, errExternal := url.Parse(app.conf.General.ExternalURL)
-		if errExternal != nil {
-			ctx.Status(http.StatusInternalServerError)
-			log.Error("Failed to parse ext url", zap.Error(errExternal))
-
-			return
-		}
-
-		ctx.SetCookie(fingerprintCookieName, "", -1, "/api",
-			parsedExternal.Hostname(), app.conf.General.Mode == ReleaseMode, true)
-
-		auth := store.PersonAuth{}
-		if errGet := app.db.GetPersonAuthByRefreshToken(ctx, fingerprint, &auth); errGet != nil {
-			responseErr(ctx, http.StatusInternalServerError, nil)
-			log.Warn("Failed to load person via fingerprint")
-
-			return
-		}
-
-		if errDelete := app.db.DeletePersonAuth(ctx, auth.PersonAuthID); errDelete != nil {
-			responseErr(ctx, http.StatusInternalServerError, nil)
-			log.Error("Failed to delete person auth on logout", zap.Error(errDelete))
-
-			return
-		}
-
-		ctx.Status(http.StatusNoContent)
-	}
 }
 
 func onOpenIDCallback(app *App) gin.HandlerFunc {
@@ -477,86 +258,6 @@ func makeTokens(ctx *gin.Context, database *store.Store, cookieKey string, sid s
 		refresh:     refreshToken,
 		fingerprint: fingerprint,
 	}, nil
-}
-
-func makeGetTokenKey(cookieKey string) func(_ *jwt.Token) (any, error) {
-	return func(_ *jwt.Token) (any, error) {
-		return []byte(cookieKey), nil
-	}
-}
-
-const fingerprintCookieName = "fingerprint"
-
-// onTokenRefresh handles generating new token pairs to access the api
-// NOTE: All error code paths must return 401 (Unauthorized).
-func onTokenRefresh(app *App) gin.HandlerFunc {
-	log := app.log.Named(runtime.FuncForPC(make([]uintptr, 10)[0]).Name())
-
-	return func(ctx *gin.Context) {
-		fingerprint, errCookie := ctx.Cookie(fingerprintCookieName)
-		if errCookie != nil {
-			ctx.AbortWithStatus(http.StatusUnauthorized)
-			log.Warn("Failed to get fingerprint cookie", zap.Error(errCookie))
-
-			return
-		}
-
-		refreshTokenString, errToken := tokenFromHeader(ctx)
-		if errToken != nil {
-			ctx.AbortWithStatus(http.StatusForbidden)
-
-			return
-		}
-
-		userClaims := userAuthClaims{}
-
-		refreshToken, errParseClaims := jwt.ParseWithClaims(refreshTokenString, &userClaims, makeGetTokenKey(app.conf.HTTP.CookieKey))
-		if errParseClaims != nil {
-			if errors.Is(errParseClaims, jwt.ErrSignatureInvalid) {
-				log.Error("jwt signature invalid!", zap.Error(errParseClaims))
-				ctx.AbortWithStatus(http.StatusUnauthorized)
-
-				return
-			}
-
-			ctx.AbortWithStatus(http.StatusUnauthorized)
-
-			return
-		}
-
-		claims, ok := refreshToken.Claims.(*userAuthClaims)
-		if !ok || !refreshToken.Valid {
-			ctx.AbortWithStatus(http.StatusUnauthorized)
-
-			return
-		}
-
-		hash := fingerprintHash(fingerprint)
-		if claims.Fingerprint != hash {
-			ctx.AbortWithStatus(http.StatusUnauthorized)
-
-			return
-		}
-
-		var auth store.PersonAuth
-		if authError := app.db.GetPersonAuthByRefreshToken(ctx, fingerprint, &auth); authError != nil {
-			ctx.AbortWithStatus(http.StatusUnauthorized)
-
-			return
-		}
-
-		tokens, errMakeToken := makeTokens(ctx, app.db, app.conf.HTTP.CookieKey, auth.SteamID, false)
-		if errMakeToken != nil {
-			ctx.AbortWithStatus(http.StatusUnauthorized)
-			log.Error("Failed to create access token pair", zap.Error(errMakeToken))
-
-			return
-		}
-
-		ctx.JSON(http.StatusOK, userToken{
-			AccessToken: tokens.access,
-		})
-	}
 }
 
 type userToken struct {
